necessary commit

+++ database/migrations/2020_06_03_125715_create_recipes_table.php
@@ -15,7 +15,7 @@ class CreateRecipesTable extends Migration
     {
         Schema::create('recipes', function (Blueprint $table) {
             $table->increments('recipesId');
-            $table->unsignedInteger('ownerId');
+            $table->unsignedBigInteger('ownerId');
             $table->string('recipesNamn');
             $table->text('recipesIngred');
             $table->text('recipesBeskrivn');
